rewrite filter test

+++ src/java/org/apache/cassandra/auth/CassandraNetworkAuthorizer.java
@@ -15,14 +15,11 @@
  * See the License for the specific language governing permissions and
  * limitations under the License.
  */
-
 package org.apache.cassandra.auth;
 
 import java.util.Set;
-
 import com.google.common.annotations.VisibleForTesting;
 import com.google.common.collect.Lists;
-
 import org.apache.cassandra.config.DatabaseDescriptor;
 import org.apache.cassandra.cql3.QueryOptions;
 import org.apache.cassandra.cql3.QueryProcessor;
@@ -37,83 +34,62 @@ import org.apache.cassandra.service.QueryState;
 import org.apache.cassandra.transport.messages.ResultMessage;
 import org.apache.cassandra.utils.ByteBufferUtil;
 
-public class CassandraNetworkAuthorizer implements INetworkAuthorizer
-{
-    private SelectStatement authorizeUserStatement = null;
+public class CassandraNetworkAuthorizer implements INetworkAuthorizer {
 
-    public void setup()
-    {
-        String query = String.format("SELECT dcs FROM %s.%s WHERE role = ?",
-                                     SchemaConstants.AUTH_KEYSPACE_NAME,
-                                     AuthKeyspace.NETWORK_PERMISSIONS);
+    public static transient org.slf4j.Logger logger_IC = org.slf4j.LoggerFactory.getLogger(CassandraNetworkAuthorizer.class);
+
+    private transient SelectStatement authorizeUserStatement = null;
+
+    public void setup() {
+        String query = String.format("SELECT dcs FROM %s.%s WHERE role = ?", SchemaConstants.AUTH_KEYSPACE_NAME, AuthKeyspace.NETWORK_PERMISSIONS);
         authorizeUserStatement = (SelectStatement) QueryProcessor.getStatement(query, ClientState.forInternalCalls());
     }
 
     @VisibleForTesting
-    ResultMessage.Rows select(SelectStatement statement, QueryOptions options)
-    {
+    ResultMessage.Rows select(SelectStatement statement, QueryOptions options) {
         return statement.execute(QueryState.forInternalCalls(), options, System.nanoTime());
     }
 
     @VisibleForTesting
-    void process(String query)
-    {
+    void process(String query) {
         QueryProcessor.process(query, ConsistencyLevel.LOCAL_ONE);
     }
 
-    private Set<String> getAuthorizedDcs(String name)
-    {
-        QueryOptions options = QueryOptions.forInternalCalls(ConsistencyLevel.LOCAL_ONE,
-                                                             Lists.newArrayList(ByteBufferUtil.bytes(name)));
-
+    private Set<String> getAuthorizedDcs(String name) {
+        QueryOptions options = QueryOptions.forInternalCalls(ConsistencyLevel.LOCAL_ONE, Lists.newArrayList(ByteBufferUtil.bytes(name)));
         ResultMessage.Rows rows = select(authorizeUserStatement, options);
         UntypedResultSet result = UntypedResultSet.create(rows.result);
         Set<String> dcs = null;
-        if (!result.isEmpty() && result.one().has("dcs"))
-        {
+        if (!result.isEmpty() && result.one().has("dcs")) {
             dcs = result.one().getFrozenSet("dcs", UTF8Type.instance);
         }
         return dcs;
     }
 
-    public DCPermissions authorize(RoleResource role)
-    {
-        if (!Roles.canLogin(role))
-        {
+    public DCPermissions authorize(RoleResource role) {
+        if (!Roles.canLogin(role)) {
             return DCPermissions.none();
         }
-        if (Roles.hasSuperuserStatus(role))
-        {
+        if (Roles.hasSuperuserStatus(role)) {
             return DCPermissions.all();
         }
-
         Set<String> dcs = getAuthorizedDcs(role.getName());
-
-        if (dcs == null || dcs.isEmpty())
-        {
+        if (dcs == null || dcs.isEmpty()) {
             return DCPermissions.all();
-        }
-        else
-        {
+        } else {
             return DCPermissions.subset(dcs);
         }
     }
 
-    private static String getSetString(DCPermissions permissions)
-    {
-        if (permissions.restrictsAccess())
-        {
+    private static String getSetString(DCPermissions permissions) {
+        if (permissions.restrictsAccess()) {
             StringBuilder builder = new StringBuilder();
             builder.append('{');
             boolean first = true;
-            for (String dc: permissions.allowedDCs())
-            {
-                if (first)
-                {
+            for (String dc : permissions.allowedDCs()) {
+                if (first) {
                     first = false;
-                }
-                else
-                {
+                } else {
                     builder.append(", ");
                 }
                 builder.append('\'');
@@ -122,36 +98,22 @@ public class CassandraNetworkAuthorizer implements INetworkAuthorizer
             }
             builder.append('}');
             return builder.toString();
-        }
-        else
-        {
+        } else {
             return "{}";
         }
     }
 
-    public void setRoleDatacenters(RoleResource role, DCPermissions permissions)
-    {
-        String query = String.format("UPDATE %s.%s SET dcs = %s WHERE role = '%s'",
-                                     SchemaConstants.AUTH_KEYSPACE_NAME,
-                                     AuthKeyspace.NETWORK_PERMISSIONS,
-                                     getSetString(permissions),
-                                     role.getName());
-
+    public void setRoleDatacenters(RoleResource role, DCPermissions permissions) {
+        String query = String.format("UPDATE %s.%s SET dcs = %s WHERE role = '%s'", SchemaConstants.AUTH_KEYSPACE_NAME, AuthKeyspace.NETWORK_PERMISSIONS, getSetString(permissions), role.getName());
         process(query);
     }
 
-    public void drop(RoleResource role)
-    {
-        String query = String.format("DELETE FROM %s.%s WHERE role = '%s'",
-                                     SchemaConstants.AUTH_KEYSPACE_NAME,
-                                     AuthKeyspace.NETWORK_PERMISSIONS,
-                                     role.getName());
-
+    public void drop(RoleResource role) {
+        String query = String.format("DELETE FROM %s.%s WHERE role = '%s'", SchemaConstants.AUTH_KEYSPACE_NAME, AuthKeyspace.NETWORK_PERMISSIONS, role.getName());
         process(query);
     }
 
-    public void validateConfiguration() throws ConfigurationException
-    {
+    public void validateConfiguration() throws ConfigurationException {
         // noop
     }
 }
